Chore/company-payment-details-fixes

+++ src/pages/shared/companyDetailView/components/CreditCardEditDelete.tsx
@@ -1,39 +1,48 @@
 import { useEffect, useState } from 'react'
 
-import { useNavigate, useParams } from 'react-router-dom'
-
 import { Button } from 'primereact/button'
 import { Chip } from 'primereact/chip'
 import { Dialog } from 'primereact/dialog'
 import { InputText } from 'primereact/inputtext'
 import { Message } from 'primereact/message'
 import { MultiSelect } from 'primereact/multiselect'
+import { Password } from 'primereact/password'
 
 import { HTLoadingLogo } from '../../../../components/shared/HTLoadingLogo'
 import { HtInputHelpText } from '../../../../components/shared/forms/HtInputHelpText'
 import { HtInputLabel } from '../../../../components/shared/forms/HtInputLabel'
 import { HtInfoTooltip } from '../../../../components/shared/general/HtInfoTooltip'
+import { type ICompanyPaymentMethod, type ICompany } from '../../../../interfaces/company'
 import { type IFacility } from '../../../../interfaces/facility'
 import { requestService } from '../../../../services/requestServiceNew'
 import { useUtils } from '../../../../store/useUtils'
 import { roleChecker } from '../../../../utils/roleChecker'
-import { type IPaymentMethod } from '../CreditCardView'
 
-export const CreditCardEditDelete = () => {
-  const { id, paymentId } = useParams()
-  const [paymentMethod, setPaymentMethod] = useState<IPaymentMethod | null>(null)
+export const CreditCardEditDelete = ({
+  companyId,
+  selectedPaymentId,
+  setSelectedPaymentId,
+  setSelectedCompanyData,
+}: {
+  companyId: string
+  selectedPaymentId: string
+  setSelectedPaymentId: (paymentId: string) => void
+  setSelectedCompanyData: (company: ICompany) => void
+}) => {
+  const [paymentMethod, setPaymentMethod] = useState<ICompanyPaymentMethod | null>(null)
   const [facilitiesByCompany, setFacilitiesByCompany] = useState<IFacility[]>([])
   const [showDialog, setShowDialog] = useState(false)
+
   const { showToast } = useUtils()
-  const navigate = useNavigate()
   const role = roleChecker()
 
   useEffect(() => {
     const fetchPaymentMethod = async () => {
       try {
-        const response = await requestService({ path: `companies/${id}/payment-method/${paymentId}` })
+        if (!selectedPaymentId) return
+        const response = await requestService({ path: `companies/${companyId}/payment-method/${selectedPaymentId}` })
         if (response.ok) {
-          const paymentMethod: IPaymentMethod = await response.json()
+          const paymentMethod: ICompanyPaymentMethod = await response.json()
           setPaymentMethod(paymentMethod)
         } else {
           console.error('Failed to fetch payment method')
@@ -44,12 +53,12 @@ export const CreditCardEditDelete = () => {
     }
 
     fetchPaymentMethod()
-  }, [id, paymentId])
+  }, [companyId, selectedPaymentId])
 
   useEffect(() => {
     const getAllFacilities = async () => {
       try {
-        const response = await requestService({ path: `facilities/company/${id}` })
+        const response = await requestService({ path: `facilities/company/${companyId}` })
         if (response.ok) {
           const fetchedFacilities: IFacility[] = await response.json()
           setFacilitiesByCompany(fetchedFacilities)
@@ -62,52 +71,58 @@ export const CreditCardEditDelete = () => {
     }
 
     getAllFacilities()
-  }, [id, paymentMethod])
+  }, [companyId])
 
   const handleDeletePaymentMethod = async () => {
     try {
+      if (paymentMethod?.is_default)
+        throw new Error('Please set another payment method as default before deleting this one.')
       const response = await requestService({
-        path: `companies/${id}/payment-method/${paymentId}`,
+        path: `companies/${companyId}/payment-method/${selectedPaymentId}`,
         method: 'DELETE',
       })
 
-      showToast({ severity: 'success', summary: 'Success', detail: 'Payment method deleted successfully' })
-      if (response.ok) {
-        await response.json()
-
-        if (role === 'client') {
-          navigate(`/client/companies/${id}/payment`)
-        } else {
-          navigate(`/admin/companies/${id}/payment`)
-        }
+      if (!response.ok) {
+        throw new Error('Failed to delete payment method')
       }
+
+      const responseData = await response.json()
+      setSelectedCompanyData(responseData)
+      setShowDialog(false)
+      setSelectedPaymentId('')
+      showToast({ severity: 'success', summary: 'Success', detail: 'Payment method deleted successfully' })
     } catch (error) {
-      console.error(error)
-      showToast({ severity: 'error', summary: 'Error', detail: 'Error deleting payment method' })
+      console.error('Error deleting payment method', error)
+      const errorMessage = error instanceof Error ? error.message : 'Error deleting payment method'
+      showToast({ severity: 'error', summary: 'Error', detail: errorMessage ?? 'Error deleting payment method' })
+      setShowDialog(false)
     }
   }
 
   const setDefaultPayment = async () => {
     try {
       const response = await requestService({
-        path: `companies/${id}/payment/${paymentId}/set-as-default`,
+        path: `companies/${companyId}/payment/${selectedPaymentId}/set-as-default`,
         method: 'POST',
       })
       if (!response.ok) {
         throw new Error('Failed to set payment method as default')
       }
-      const responseData = await response.json()
-      setPaymentMethod(responseData)
+      const responseData: ICompany = await response.json()
+      setSelectedCompanyData(responseData)
+      setSelectedPaymentId('')
+      showToast({ severity: 'success', summary: 'Success', detail: 'Payment method set as default' })
     } catch (error) {
       console.error('Error setting default payment method: ', error)
+      showToast({ severity: 'error', summary: 'Error', detail: 'Error setting default payment method' })
     }
   }
 
   const renderDialogFooter = () => {
     return (
       <div>
-        <Button label="No" onClick={() => setShowDialog(false)} className="p-button-text" />
-        <Button label="Yes" onClick={handleDeletePaymentMethod} />
+        <Button severity="secondary" label="No" text outlined onClick={() => setShowDialog(false)} />
+        <Button severity="danger" label="Yes" onClick={handleDeletePaymentMethod} />
       </div>
     )
   }
@@ -141,13 +156,13 @@ export const CreditCardEditDelete = () => {
                     <HtInfoTooltip message="Reference name for the card in the application.">
                       <HtInputLabel htmlFor="card_name" asterisk labelText="Card Name" />
                     </HtInfoTooltip>
-                    <InputText id="card_name" value={paymentMethod.payment_info.card_name} className="mt-2" readOnly />
+                    <InputText id="card_name" value={paymentMethod.payment_info.card_name} className="mt-2" disabled />
                   </div>
                   <div className="sm:col-span-3">
                     <HtInfoTooltip message="Provide billing address.">
                       <HtInputLabel htmlFor="address" asterisk labelText="Billing Address" />
                     </HtInfoTooltip>
-                    <InputText id="address" value={paymentMethod.payment_info.address} className="mt-2" readOnly />
+                    <InputText id="address" value={paymentMethod.payment_info.address} className="mt-2" disabled />
                   </div>
                 </>
               ) : null}
@@ -162,7 +177,7 @@ export const CreditCardEditDelete = () => {
                       id="ach_account_name"
                       value={paymentMethod.payment_info.ach_account_name}
                       className="mt-2"
-                      readOnly
+                      disabled
                     />
                   </div>
                   <div className="sm:col-span-3">
@@ -173,7 +188,7 @@ export const CreditCardEditDelete = () => {
                       id="ach_bank_name"
                       value={paymentMethod.payment_info.ach_bank_name}
                       className="mt-2"
-                      readOnly
+                      disabled
                     />
                   </div>
 
@@ -181,11 +196,15 @@ export const CreditCardEditDelete = () => {
                     <HtInfoTooltip message="The account number is the unique identifier for the bank account.">
                       <HtInputLabel htmlFor="ach_account_number" asterisk labelText="Account Number" />
                     </HtInfoTooltip>
-                    <InputText
+                    <Password
                       id="ach_account_number"
                       value={paymentMethod.payment_info.ach_account_number}
                       className="mt-2"
-                      readOnly
+                      disabled
+                      toggleMask
+                      pt={{
+                        panel: { className: 'hidden' },
+                      }}
                     />
                   </div>
 
@@ -193,11 +212,15 @@ export const CreditCardEditDelete = () => {
                     <HtInfoTooltip message="The routing number is a nine-digit number that identifies the bank.">
                       <HtInputLabel htmlFor="routing_number" asterisk labelText="Routing Number" />
                     </HtInfoTooltip>
-                    <InputText
+                    <Password
                       id="ach_routing_number"
                       value={paymentMethod.payment_info.ach_routing_number}
                       className="mt-2"
-                      readOnly
+                      disabled
+                      toggleMask
+                      pt={{
+                        panel: { className: 'hidden' },
+                      }}
                     />
                   </div>
                 </>
@@ -225,39 +248,44 @@ export const CreditCardEditDelete = () => {
                 />
               </div>
 
-              {paymentMethod.is_default === true ? (
-                <Chip className="ml-2" label="Default Payment Method" icon="pi pi-check" />
-              ) : (
-                <Button label="Set as Default" severity="secondary" text raised onClick={setDefaultPayment} />
-              )}
+              <div className="sm:col-span-3">
+                {paymentMethod.is_default === true ? (
+                  <Chip className="ml-2" label="Default Payment Method" icon="pi pi-check" />
+                ) : (
+                  <Button label="Set as Default" severity="secondary" raised onClick={setDefaultPayment} />
+                )}
+              </div>
             </div>
           </div>
         </div>
         <div className="mt-6 flex items-center justify-end gap-x-6" />
       </div>
-      <div className="grid grid-cols-1 gap-x-8 gap-y-4  border-gray-900/10 pb-12 sm:gap-y-10 md:grid-cols-3">
-        <h2 className="text-base font-semibold leading-7 text-gray-900">Delete Payment Method</h2>
-        <div className="grid max-w-2xl grid-cols-1 gap-x-6 gap-y-6 sm:grid-cols-1 md:col-span-2">
-          <div className="flex flex-col items-center sm:col-span-3">
-            <Dialog
-              visible={showDialog}
-              style={{ width: '450px' }}
-              header="Confirm"
-              modal
-              footer={renderDialogFooter()}
-              onHide={() => setShowDialog(false)}>
-              Are you sure you want to delete this payment method?
-            </Dialog>
-            <Message
-              severity="error"
-              text="This will delete this payment method from the company. This action cannot be undone."
-            />
-            <Button severity="danger" className="mt-2" onClick={() => setShowDialog(true)}>
-              Delete Payment Method
-            </Button>
+      {role === 'admin' ? (
+        <div className="grid grid-cols-1 gap-x-8 gap-y-4  border-gray-900/10 pb-12 sm:gap-y-10 md:grid-cols-3">
+          <h2 className="text-base font-semibold leading-7 text-gray-900">Delete Payment Method</h2>
+          <div className="grid max-w-2xl grid-cols-1 gap-x-6 gap-y-6 sm:grid-cols-1 md:col-span-2">
+            <div className="flex flex-col items-center sm:col-span-3">
+              <Dialog
+                visible={showDialog}
+                style={{ width: '450px' }}
+                header="Confirm"
+                modal
+                draggable={false}
+                footer={renderDialogFooter()}
+                onHide={() => setShowDialog(false)}>
+                Are you sure you want to delete this payment method?
+              </Dialog>
+              <Message
+                severity="error"
+                text="This will delete this payment method from the company. This action cannot be undone."
+              />
+              <Button severity="danger" className="mt-2" onClick={() => setShowDialog(true)}>
+                Delete Payment Method
+              </Button>
+            </div>
           </div>
         </div>
-      </div>
+      ) : null}
     </>
   ) : (
     <HTLoadingLogo />
